requirements.txt
astunparse==1.6.3
backports.tarfile==1.2.0
cartesia==1.3.1
deepgram-sdk==3.7.7
flask-cors==5.0.0
flask-socketio==5.5.0
gast==0.6.0
google-pasta==0.2.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
ipykernel==6.29.5
jaraco.collections==5.1.0
keras==3.5.0
openai==1.59.3
opt-einsum==3.3.0
pandas==2.2.3
pip-chill==1.0.3
pipreqs==0.5.0
praw==7.7.1
pyaudio==0.2.14
pydub==0.25.1
pymupdf==1.25.1
python-dotenv==1.0.1
tensorboard==2.17.1
tinycss2==1.4.0
tomli==2.0.1
twilio==9.4.1
websocket==0.2.1
